chore(core): use TS for next config
