+++ .changeset/shaggy-penguins-act.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Use Typescript on Next Config
